refactor(SaveFileViewController): Split the class for better readability and further development
