+++ kDrive/UI/Controller/Files/Save File/SaveFileViewController+SelectFolderDelegate.swift
@@ -1,0 +1,34 @@
+/*
+ Infomaniak kDrive - iOS App
+ Copyright (C) 2024 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import kDriveCore
+import UIKit
+
+// MARK: - SelectFolderDelegate
+
+extension SaveFileViewController: SelectFolderDelegate {
+    func didSelectFolder(_ folder: File) {
+        if folder.id == DriveFileManager.constants.rootID {
+            selectedDirectory = selectedDriveFileManager?.getCachedRootFile()
+        } else {
+            selectedDirectory = folder
+        }
+        updateButton()
+        tableView.reloadData()
+    }
+}
