+++ kDrive/UI/Controller/Files/Save File/SaveFileViewController.swift
@@ -16,7 +16,6 @@
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import CocoaLumberjackSwift
 import InfomaniakCore
 import InfomaniakCoreUIKit
 import InfomaniakDI
@@ -29,6 +28,16 @@ class SaveFileViewController: UIViewController {
     @LazyInjectService var accountManager: AccountManageable
     @LazyInjectService var fileImportHelper: FileImportHelper
     @LazyInjectService var appContextService: AppContextServiceable
+
+    private var originalDriveId: Int = {
+        @InjectService var accountManager: AccountManageable
+        return accountManager.currentDriveId
+    }()
+
+    private var originalUserId: Int = {
+        @InjectService var accountManager: AccountManageable
+        return accountManager.currentUserId
+    }()
 
     enum SaveFileSection {
         case alert
@@ -46,16 +55,6 @@ class SaveFileViewController: UIViewController {
     }
 
     var sections: [SaveFileSection] = [.fileName, .driveSelection, .directorySelection]
-
-    private var originalDriveId: Int = {
-        @InjectService var accountManager: AccountManageable
-        return accountManager.currentDriveId
-    }()
-
-    private var originalUserId: Int = {
-        @InjectService var accountManager: AccountManageable
-        return accountManager.currentUserId
-    }()
 
     var selectedDriveFileManager: DriveFileManager?
     var selectedDirectory: File?
@@ -92,9 +91,9 @@ class SaveFileViewController: UIViewController {
         return false
     }
 
-    private var errorCount = 0
-    private var importProgress: Progress?
-    private var enableButton = false {
+    var errorCount = 0
+    var importProgress: Progress?
+    var enableButton = false {
         didSet {
             guard let footer = tableView.footerView(forSection: tableView.numberOfSections - 1) as? FooterButtonView else {
                 return
@@ -103,7 +102,7 @@ class SaveFileViewController: UIViewController {
         }
     }
 
-    private var importInProgress: Bool {
+    var importInProgress: Bool {
         if let progress = importProgress {
             return progress.fractionCompleted < 1
         } else {
@@ -159,6 +158,88 @@ class SaveFileViewController: UIViewController {
         )
     }
 
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        MatomoUtils.track(view: [MatomoUtils.Views.save.displayName, "SaveFile"])
+    }
+
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        navigationController?.setInfomaniakAppearanceNavigationBar()
+        tableView.reloadData()
+    }
+
+    deinit {
+        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
+        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
+    }
+
+    @objc func keyboardWillShow(_ notification: Notification) {
+        if let keyboardSize = (notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue)?.cgRectValue {
+            tableView.contentInset.bottom = keyboardSize.height
+
+            UIView.animate(withDuration: 0.1) {
+                self.view.layoutIfNeeded()
+            }
+        }
+    }
+
+    @objc func keyboardWillHide(_ notification: Notification) {
+        tableView.contentInset.bottom = 0
+        UIView.animate(withDuration: 0.1) {
+            self.view.layoutIfNeeded()
+        }
+    }
+
+    @IBAction func close(_ sender: Any) {
+        importProgress?.cancel()
+        dismiss(animated: true)
+        if let extensionContext {
+            extensionContext.completeRequest(returningItems: nil, completionHandler: nil)
+        }
+    }
+
+    func setAssetIdentifiers() {
+        guard let assetIdentifiers else { return }
+        sections = [.importing]
+        importProgress = fileImportHelper.importAssets(
+            assetIdentifiers,
+            userPreferredPhotoFormat: userPreferredPhotoFormat
+        ) { [weak self] importedFiles, errorCount in
+            guard let self else {
+                return
+            }
+
+            items = importedFiles
+            self.errorCount = errorCount
+            Task { @MainActor in
+                self.updateTableViewAfterImport()
+            }
+        }
+    }
+
+    func setItemProviders() {
+        guard let itemProviders else { return }
+        sections = [.importing]
+        importProgress = fileImportHelper
+            .importItems(itemProviders,
+                         userPreferredPhotoFormat: userPreferredPhotoFormat) { [weak self] importedFiles, errorCount in
+                guard let self else {
+                    return
+                }
+
+                items = importedFiles
+                self.errorCount = errorCount
+                Task { @MainActor in
+                    self.updateTableViewAfterImport()
+                }
+            }
+    }
+
+    func updateButton() {
+        enableButton = selectedDirectory != nil && items.allSatisfy { !$0.name.isEmpty } && !items.isEmpty && !importInProgress
+    }
+
     func getBestDirectory() -> File? {
         if lastSelectedDirectory?.driveId == selectedDriveFileManager?.drive.id {
             return lastSelectedDirectory
@@ -187,28 +268,6 @@ class SaveFileViewController: UIViewController {
         return firstAvailableSharedDriveDirectory?.freezeIfNeeded()
     }
 
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        MatomoUtils.track(view: [MatomoUtils.Views.save.displayName, "SaveFile"])
-    }
-
-    @objc func keyboardWillShow(_ notification: Notification) {
-        if let keyboardSize = (notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue)?.cgRectValue {
-            tableView.contentInset.bottom = keyboardSize.height
-
-            UIView.animate(withDuration: 0.1) {
-                self.view.layoutIfNeeded()
-            }
-        }
-    }
-
-    @objc func keyboardWillHide(_ notification: Notification) {
-        tableView.contentInset.bottom = 0
-        UIView.animate(withDuration: 0.1) {
-            self.view.layoutIfNeeded()
-        }
-    }
-
     func dismiss(animated: Bool, clean: Bool = true, completion: (() -> Void)? = nil) {
         Task {
             // Cleanup file that were duplicated to appGroup on extension mode
@@ -220,52 +279,6 @@ class SaveFileViewController: UIViewController {
 
             navigationController?.dismiss(animated: animated, completion: completion)
         }
-    }
-
-    deinit {
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-
-    private func setAssetIdentifiers() {
-        guard let assetIdentifiers else { return }
-        sections = [.importing]
-        importProgress = fileImportHelper.importAssets(
-            assetIdentifiers,
-            userPreferredPhotoFormat: userPreferredPhotoFormat
-        ) { [weak self] importedFiles, errorCount in
-            guard let self else {
-                return
-            }
-
-            items = importedFiles
-            self.errorCount = errorCount
-            Task { @MainActor in
-                self.updateTableViewAfterImport()
-            }
-        }
-    }
-
-    private func setItemProviders() {
-        guard let itemProviders else { return }
-        sections = [.importing]
-        importProgress = fileImportHelper
-            .importItems(itemProviders,
-                         userPreferredPhotoFormat: userPreferredPhotoFormat) { [weak self] importedFiles, errorCount in
-                guard let self else {
-                    return
-                }
-
-                items = importedFiles
-                self.errorCount = errorCount
-                Task { @MainActor in
-                    self.updateTableViewAfterImport()
-                }
-            }
-    }
-
-    private func updateButton() {
-        enableButton = selectedDirectory != nil && items.allSatisfy { !$0.name.isEmpty } && !items.isEmpty && !importInProgress
     }
 
     private func updateTableViewAfterImport() {
@@ -298,12 +311,6 @@ class SaveFileViewController: UIViewController {
         }
     }
 
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        navigationController?.setInfomaniakAppearanceNavigationBar()
-        tableView.reloadData()
-    }
-
     class func instantiate(driveFileManager: DriveFileManager?) -> SaveFileViewController {
         let viewController = Storyboard.saveFile
             .instantiateViewController(withIdentifier: "SaveFileViewController") as! SaveFileViewController
@@ -320,209 +327,5 @@ class SaveFileViewController: UIViewController {
         let navigationController = TitleSizeAdjustingNavigationController(rootViewController: saveViewController)
         navigationController.navigationBar.prefersLargeTitles = true
         return navigationController
-    }
-
-    @IBAction func close(_ sender: Any) {
-        importProgress?.cancel()
-        dismiss(animated: true)
-        if let extensionContext {
-            extensionContext.completeRequest(returningItems: nil, completionHandler: nil)
-        }
-    }
-}
-
-// MARK: - UITableViewDataSource
-
-extension SaveFileViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        let section = sections[section]
-        if section == .fileName {
-            return items.count
-        } else {
-            return 1
-        }
-    }
-
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return sections.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        switch sections[indexPath.section] {
-        case .alert:
-            let cell = tableView.dequeueReusableCell(type: AlertTableViewCell.self, for: indexPath)
-            cell.configure(with: .warning, message: KDriveResourcesStrings.Localizable.snackBarUploadError(errorCount))
-            return cell
-        case .fileName:
-            let item = items[indexPath.row]
-            if items.count > 1 {
-                let cell = tableView.dequeueReusableCell(type: UploadTableViewCell.self, for: indexPath)
-                cell.initWithPositionAndShadow(
-                    isFirst: indexPath.row == 0,
-                    isLast: indexPath.row == self.tableView(tableView, numberOfRowsInSection: indexPath.section) - 1
-                )
-                cell.configureWith(importedFile: item)
-                return cell
-            } else {
-                let cell = tableView.dequeueReusableCell(type: FileNameTableViewCell.self, for: indexPath)
-                cell.textField.text = item.name
-                cell.textDidChange = { [weak self] text in
-                    guard let self else { return }
-                    item.name = text ?? KDriveResourcesStrings.Localizable.allUntitledFileName
-                    if let text, !text.isEmpty {
-                        updateButton()
-                    } else {
-                        enableButton = false
-                    }
-                }
-                return cell
-            }
-        case .driveSelection:
-            let cell = tableView.dequeueReusableCell(type: LocationTableViewCell.self, for: indexPath)
-            cell.initWithPositionAndShadow(isFirst: true, isLast: true)
-            cell.configure(with: selectedDriveFileManager?.drive)
-            return cell
-        case .directorySelection:
-            let cell = tableView.dequeueReusableCell(type: LocationTableViewCell.self, for: indexPath)
-            cell.initWithPositionAndShadow(isFirst: true, isLast: true)
-            cell.configure(with: selectedDirectory, drive: selectedDriveFileManager!.drive)
-            return cell
-        case .photoFormatOption:
-            let cell = tableView.dequeueReusableCell(type: PhotoFormatTableViewCell.self, for: indexPath)
-            cell.initWithPositionAndShadow(isFirst: true, isLast: true)
-            cell.configure(with: userPreferredPhotoFormat)
-            return cell
-        case .importing:
-            let cell = tableView.dequeueReusableCell(type: ImportingTableViewCell.self, for: indexPath)
-            cell.importationProgressView.observedProgress = importProgress
-            return cell
-        default:
-            fatalError("Not supported by this datasource")
-        }
-    }
-
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        switch sections[section] {
-        case .fileName:
-            return HomeTitleView.instantiate(title: "")
-        case .driveSelection:
-            return HomeTitleView.instantiate(title: "kDrive")
-        case .directorySelection:
-            return HomeTitleView.instantiate(title: KDriveResourcesStrings.Localizable.allPathTitle)
-        case .photoFormatOption:
-            return HomeTitleView.instantiate(title: KDriveResourcesStrings.Localizable.photoFormatTitle)
-        default:
-            return nil
-        }
-    }
-
-    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        if section == tableView.numberOfSections - 1 && !importInProgress {
-            return 124
-        }
-        return 32
-    }
-
-    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        if section == tableView.numberOfSections - 1 && !importInProgress {
-            let view = FooterButtonView.instantiate(title: KDriveResourcesStrings.Localizable.buttonSave)
-            view.delegate = self
-            view.footerButton.isEnabled = enableButton
-            return view
-        }
-        return nil
-    }
-}
-
-// MARK: - UITableViewDelegate
-
-extension SaveFileViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-        switch sections[indexPath.section] {
-        case .fileName:
-            let item = items[indexPath.row]
-            if items.count > 1 {
-                let alert = AlertFieldViewController(
-                    title: KDriveResourcesStrings.Localizable.buttonRename,
-                    placeholder: KDriveResourcesStrings.Localizable.hintInputFileName,
-                    text: item.name,
-                    action: KDriveResourcesStrings.Localizable.buttonSave,
-                    loading: false
-                ) { newName in
-                    item.name = newName
-                    tableView.reloadRows(at: [indexPath], with: .automatic)
-                }
-                alert.textFieldConfiguration = .fileNameConfiguration
-                alert.textFieldConfiguration.selectedRange = item.name
-                    .startIndex ..< (item.name.lastIndex { $0 == "." } ?? item.name.endIndex)
-                present(alert, animated: true)
-            }
-        case .driveSelection:
-            let selectDriveViewController = SelectDriveViewController.instantiate()
-            selectDriveViewController.selectedDrive = selectedDriveFileManager?.drive
-            selectDriveViewController.delegate = self
-            navigationController?.pushViewController(selectDriveViewController, animated: true)
-        case .directorySelection:
-            guard let driveFileManager = selectedDriveFileManager else { return }
-            let selectFolderNavigationController = SelectFolderViewController.instantiateInNavigationController(
-                driveFileManager: driveFileManager,
-                startDirectory: selectedDirectory,
-                delegate: self
-            )
-            present(selectFolderNavigationController, animated: true)
-        case .photoFormatOption:
-            let selectPhotoFormatViewController = SelectPhotoFormatViewController
-                .instantiate(selectedFormat: userPreferredPhotoFormat)
-            selectPhotoFormatViewController.delegate = self
-            navigationController?.pushViewController(selectPhotoFormatViewController, animated: true)
-        default:
-            break
-        }
-    }
-}
-
-// MARK: - SelectFolderDelegate
-
-extension SaveFileViewController: SelectFolderDelegate {
-    func didSelectFolder(_ folder: File) {
-        if folder.id == DriveFileManager.constants.rootID {
-            selectedDirectory = selectedDriveFileManager?.getCachedRootFile()
-        } else {
-            selectedDirectory = folder
-        }
-        updateButton()
-        tableView.reloadData()
-    }
-}
-
-// MARK: - SelectDriveDelegate
-
-extension SaveFileViewController: SelectDriveDelegate {
-    func didSelectDrive(_ drive: Drive) {
-        if let selectedDriveFileManager = accountManager.getDriveFileManager(for: drive.id, userId: drive.userId) {
-            self.selectedDriveFileManager = selectedDriveFileManager
-            selectedDirectory = getBestDirectory()
-            sections = [.fileName, .driveSelection, .directorySelection]
-            if itemProvidersContainHeicPhotos {
-                sections.append(.photoFormatOption)
-            }
-        }
-        updateButton()
-    }
-}
-
-// MARK: - SelectPhotoFormatDelegate
-
-extension SaveFileViewController: SelectPhotoFormatDelegate {
-    func didSelectPhotoFormat(_ format: PhotoFileFormat) {
-        if userPreferredPhotoFormat != format {
-            userPreferredPhotoFormat = format
-            if itemProviders?.isEmpty == false {
-                setItemProviders()
-            } else {
-                setAssetIdentifiers()
-            }
-        }
     }
 }
